+++ src/main/java/pro/sky/finalprojectsky/repository/AdsRepository.java
@@ -2,10 +2,14 @@ package pro.sky.finalprojectsky.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import pro.sky.finalprojectsky.entity.Ads;
+
 import java.util.List;
 
-public interface AdsRepository extends JpaRepository<Ads, Integer> {
+/**
+ * Repository AdsRepository (advertisement/объявление).
+ */
+public interface AdsRepository extends JpaRepository<Ads, Long> {
 
-    List<Ads> findAllByAuthorId(Integer id);
+    List<Ads> findAllByAuthorId(long id);
 
 }
