+++ src/main/java/pro/sky/finalprojectsky/repository/ImageRepository.java
@@ -3,6 +3,10 @@ package pro.sky.finalprojectsky.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pro.sky.finalprojectsky.entity.Image;
 
-public interface ImageRepository extends JpaRepository<Image, Integer> {
-    Image findByAdsId(Integer adsId);
+/**
+ * Repository ImageRepository (advertisement image/изображение в объявлениях).
+ */
+public interface ImageRepository extends JpaRepository<Image, Long> {
+    Image findByAdsId(long adsId);
+
 }
