+++ src/main/java/pro/sky/finalprojectsky/repository/UserRepository.java
@@ -2,9 +2,17 @@ package pro.sky.finalprojectsky.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import pro.sky.finalprojectsky.entity.User;
+
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+/**
+ * Repository UserRepository (users/пользователь).
+ */
+public interface UserRepository extends JpaRepository<User, Long> {
+
     Optional<User> findByEmail(String email);
+
     boolean existsByEmail(String email);
+
+
 }
